BACKPORT: gpio_emul: Fix bug with input & output pins
Change gpio_emul_input_set_masked_pend to log when it returns an error.
Changed all calls to gpio_emul_input_set_masked* to check the return
value.

Changed the function gpio_emul_input_set_masked_pend to apply the mask
to the provided values and not return an error if there are excess
values. In almost every call to that function, the value wasn't being
limited to bits set in the mask, so they were failing if there was more
than one gpio configured for INPUT|OUTPUT with a value of 1.

Fixes #40646

BUG=b:207029014

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: I6d6e6a2eba6374113c4fb4a34098548144d2a747
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3308901
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>