Bluetooth: LC3: Addition of liblc3codec for LE-Audio as a module
Initial addition of the LC3 codec as a module. Usage of the module
will be seperate commits.

Even though this codec is a generic audio codec, it is tagged with
other purposes is a violation of the granted rights for the codec.

Bluetooth: as it can only be used for LE-Audio. Using the codec for
Signed-off-by: Casper Bonde <casper_bonde@bose.com>
GitOrigin-RevId: d03d2d854b49f32da4322ca1d682d52ce85269ee
Change-Id: Ib46ad9ece03f541ece7066f374def0fe15cf7dd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3562742
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>