tests: kernel: threads: Correct test assertion in thread_apis test
Thread APIs test for k_busy_wait incorrectly asserted that a delay of
100 us should produce a delay in cycles less than or equal to to 100 cycles
of the hardware clock. Since most hardware clocks are fast, this assertion
was valid, but it does not test for the actual delay.

Fix the assertion to verify that a delay of 100 us produces a delay in
cycles less than or equal to 100 us worth of hardware clock cycles.

(cherry picked from commit 369a54514fbca30254fe319f22409b1b75c19fc3)

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>
GitOrigin-RevId: 369a54514fbca30254fe319f22409b1b75c19fc3
Change-Id: Ic13f17f9e0386c595c4d554c4750b8c300875c32
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3338820
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>