soc/intel_adsp: Align .rodata with .data in bootloader
This seems to be a mistake in rimage: it wants the text and data
segments of the output module to be page-aligned, but it assumes
.rodata is part of "data" and not "text".  So this reorders the
segments to make that happen.

Note that the page alignment is entirely artificial.  Nothing is
interpreting the segment boundaries rimage is enforcing except for the
code in the bootloader itself, which doesn't care.

(cherry picked from commit f1e941acd4e25399b42e5cf1667fed171b20d3ea)

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: f1e941acd4e25399b42e5cf1667fed171b20d3ea
Change-Id: I038490e8def7439e8b8865f49debfe4473e191d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3341732
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>