arch: cortex_m: Fix dwt cyccnt assert
Fix the assert that checks for existence of a cycle counter.

The field is named NO CYCCNT, so when it is 1, there is no cycle
counter. But we are asserting the opposite.

(cherry picked from commit 1f87642f083a7cee0aedad4bc349684ef87b1e71)

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
GitOrigin-RevId: 1f87642f083a7cee0aedad4bc349684ef87b1e71
Change-Id: Ibe728e9c67d9b2fff963e052ab1bcc11d6a04264
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3329994
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>