kernel: update z_sched_thread_usage()
This commit does two things to the z_sched_thread_usage(). First,
it updates the API so that it accepts a pointer to the runtime
stats instead of simply returning the usage cycles. This gives it
the flexibility to retrieve additional statistics in the future.

Second, the runtime stats are only updated if the specified thread
is the current thread running on the current core.

(cherry picked from commit 5deaffb2eee2e39618bc3c50ce2f428c6581477a)

Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>
GitOrigin-RevId: 5deaffb2eee2e39618bc3c50ce2f428c6581477a
Change-Id: Ice19ec872023693ef89ab4dce4dcaba01534f9ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3380737
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>