net: lwm2m: Improve content writers error handling in engine
Verify the restult of the content writer functions and return an error
if content writer fails to read/write field in the message.

This solves an issue when for example malformed packet was sent to the
server if the payload did not fit in the message.

(cherry picked from commit cb6c5f37ca7fbd84cf0e057db7b24bd6675fd64c)

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
GitOrigin-RevId: cb6c5f37ca7fbd84cf0e057db7b24bd6675fd64c
Change-Id: Ic7c08aa114b672ade16f6d9bd2e93e6373535dcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3431230
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>