bluetooth: ots: validate OTS procedure sizes
Validate the OACP and OLCP procedure sizes as the procedure is parsed.
This is in contrast with the current implementation which first parses
the procedure and then validates its size and is prone to accessing
invalid memory when the procedure is malformed.

(cherry picked from commit 45a693a2a92412c8677cae9c3111bb4369272f42)

Signed-off-by: Abe Kohandel <abe.kohandel@gmail.com>
GitOrigin-RevId: 45a693a2a92412c8677cae9c3111bb4369272f42
Change-Id: I86702eab08381807b676e044b0ed4f815a87c88a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3386242
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>