soc/intel_adsp: Use the sys_winstream protocol in adsplog.py
This is the matching commit to the previous one that swaps the
protocol used for window logging for sys_winstream.

The advantage is especially clear for the reader.  The old protocol
was complicated and race-prone, requiring whole-buffer reads for
reliability.  The new one is tiny and fast.

(cherry picked from commit 8cb5bf152686ad9060162ce5b0215384536e9585)

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: 8cb5bf152686ad9060162ce5b0215384536e9585
Change-Id: Ifade826c399be7441a75407a3bef02b478b2f010
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3387562
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>