Bluetooth: ATT: Change prep_queue to slist instead of fifo
Change the type of the prep_queue to a linked list
instead of a fifo. The main reason for this, is
that we can then iterate over the content in it,
without popping it.

(cherry picked from commit a1ce2d4c83f361c7160fc15f703097ab2d20e5cf)

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: a1ce2d4c83f361c7160fc15f703097ab2d20e5cf
Change-Id: I9737c68991b2997d1c42a5e2b3fb01136681585e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3367607
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>