kernel/smp: Fix races in SMP initialization
This has bitrotten a bit.  Early implementations had a synchronous
arch_start_cpu(), but then we started allowing that to be an async
operation.  But that means that CPU start now becomes surprisingly
reentrant to the arch layer (cpu 0 can get a call to start cpu 2 while
cpu 1's initialization code is still running).  That's just error
prone; we never documented the requirements cleanly (the window is
very small, but not so small to a slow simulator!).

Add an extra flag so we don't issue the next start until the last is
out of the arch layer and running in smp_init_top().

(cherry picked from commit 3457118540f77a27bc7c2f1fb3331474ee9948b2)

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: 3457118540f77a27bc7c2f1fb3331474ee9948b2
Change-Id: Iee5adab761f7e75483c2226671d061ac0555d669
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3380831
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>