cmake: bintools: gnu: Suppress GDB check console outputs
The GDB check routine added in the PR #38749 does not suppress the
console outputs (stdout and stderr) and may print out a misleading
error message during a CMake configuration when the required version
of Python is not available on the system:

  arm-zephyr-eabi-gdb: error while loading shared libraries:
  libpython3.8.so.1.0: cannot open shared object file: No such file or
  directory

This commit adds the `OUTPUT_QUIET` and `ERROR_QUIET` options when
executing the GDB process so that the console outputs during the GDB
executable validation are not displayed to the user.

In addition, this commit removes the unused `GDB_PY_NO_PY` standard
output redirection variable since it is unnecessary.

(cherry picked from commit 3181331586f1fdde1d1ac3139ebdfd188baabbfd)

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: 3181331586f1fdde1d1ac3139ebdfd188baabbfd
Change-Id: I22b6b73508c4c3656e1c0bac1d9a3ed70ee01f70
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3380743
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>