drivers: flash: nrf_qspi_nor: Fix setting of the base clock divider
On nRF53 Series SoCs, the highest SCK frequencies can only be achieved
when the HFCLK192M clock divider is changed from the default /4 setting.
Such change results in increased power consumption, so the divider needs
to be changed only for periods when it is actually necessary.
This commit modifies the driver behavior so that it changes the divider
only when a QSPI bus operation is performed.
However, when XIP accesses to the flash chip are also used, it may be
needed to keep the divider changed even when the driver is idle so that
the XIP access speed is not reduced, hence a custom API function that
allows forcing this is introduced for the driver.

(cherry picked from commit aa3b86600db65873cd83a63de873b3414a9a564d)

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: aa3b86600db65873cd83a63de873b3414a9a564d
Change-Id: I091e1d725a6785b09e7c39ff4393a6b5d5120d25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3425611
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>