cmake: support ELF image adjustment
This commit adds support for adjust the addresses of the final image.
This is useful when the image is to be flashed at a location different
from the LMA address encoded in the ELF file by the linker.

An example use-case is multicore systems where core A might load image
from a flash partition into RAM in order for core B to execute and load,
but where the image itself is build with the RAM addresses as LMA.

It updates the zephyr_image_info.h header with information of adjustment
value.

(cherry picked from commit d51a67b7c098c8bb5479a7b87f05dc40fb76e065)

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>
GitOrigin-RevId: d51a67b7c098c8bb5479a7b87f05dc40fb76e065
Change-Id: I609e91dd90d4a922be631b40caa182e398fa864c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3410999
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>