bluetooth: ots: notify application prior to object name change
This change augments the application object name change notification to
occur prior to the actual name change done by the OTS layer.

Notifying prior to the actual name change makes it possible to inform the
application of the current name of the object as well as the new name.

Signed-off-by: Abe Kohandel <abe.kohandel@gmail.com>
GitOrigin-RevId: f1e75dea501630343bf259cdac3070baa238724b
Change-Id: I3e7c24cbf76a8eea7d9299aac010087638798c65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3579054
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>