dma: Add host related dma transfer directions
DMA between a host os and processor (such as an x86 processor running
linux) and vice versa is used and supported by Intel's HDA DMA
Controller. This provides a method to transfer data, using hardware,
between the host and the local processor running Zephyr. The enums
added here are used to extend the dma_status direction enum
which provides information back to the DMA user and driver the
direction of the DMA transfer for a particular channel. This can
then be used, where needed, to add logic around the direction of the
transfer.

In addition to adding the two directions it adds enums to account for
cases where additional DMA directions that are IP and driver specific
may exist.

(cherry picked from commit 669c4d17387928f2010791bdeaa2a07ac39565b0)

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>
GitOrigin-RevId: 669c4d17387928f2010791bdeaa2a07ac39565b0
Change-Id: I5c032bc167a2afe6c5f9604b4076c342127384f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3404485
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>