kernel: workq: Fix type errors in delayable work handlers
A common pattern here was to take the work item as the subfield of a
containing object. But the contained field is not a k_work, it's a
k_work_delayable.

Things were working only because the work field was first, so the
pointers had the same value. Do things right and fix things to
produce correct code if/when that field ever moves within delayable.

(cherry picked from commit 731241f8d049924591cf4f46dda26ddd0782d8c7)

Signed-off-by: Yong Cong Sin <yongcong.sin@gmail.com>
GitOrigin-RevId: 731241f8d049924591cf4f46dda26ddd0782d8c7
Change-Id: Ie2d27a33d0c709d38d7d9b7fd3c050c4a0e69965
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3436254
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>