ztest: run before function in test thread
A lot of tests need to be able to get their current tid and do some
action with it. It makes sense for the `before` function/rule to be
able to run in the same thread as the test. Note that the `after`
function does not run in the same thread because we need to guarantee
that it will run.

(cherry picked from commit 9c0b970bc257f995f959a4d58dcef0f708a9a937)

Signed-off-by: Yuval Peress <peress@google.com>
GitOrigin-RevId: 9c0b970bc257f995f959a4d58dcef0f708a9a937
Change-Id: Ieead1a7d16d2e05410fc3b274d6323c497e248f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3406638
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>